Move perform validate to validate.py
